+++ objc2/src/declare/ivar_drop.rs
@@ -278,8 +278,7 @@ mod tests {
     fn test_alloc_dealloc() {
         let expected = ThreadTestData::current();
 
-        let obj: Id<Allocated<IvarTester>, Owned> =
-            unsafe { msg_send_id![IvarTester::class(), alloc] };
+        let obj: Allocated<IvarTester> = unsafe { msg_send_id![IvarTester::class(), alloc] };
         expected.assert_current();
 
         drop(obj);
@@ -315,8 +314,7 @@ mod tests {
     #[cfg_attr(not(debug_assertions), ignore = "only panics in debug mode")]
     #[should_panic = "an Id in instance variables must always be initialized before use"]
     fn test_init_invalid_ref() {
-        let obj: Id<IvarTester, Owned> =
-            unsafe { msg_send_id![msg_send_id![IvarTester::class(), alloc], initInvalid] };
+        let obj: Id<IvarTester, Owned> = unsafe { msg_send_id![IvarTester::alloc(), initInvalid] };
 
         std::println!("{:?}", obj.ivar1);
     }
@@ -326,7 +324,7 @@ mod tests {
     #[should_panic = "an Id in instance variables must always be initialized before use"]
     fn test_init_invalid_mut() {
         let mut obj: Id<IvarTester, Owned> =
-            unsafe { msg_send_id![msg_send_id![IvarTester::class(), alloc], initInvalid] };
+            unsafe { msg_send_id![IvarTester::alloc(), initInvalid] };
 
         *obj.ivar1 = RcTestObject::new().into();
     }